Made needed changes/corrections outlined in pr 22 on github. Changed some type/interface names, updated the changeset,  changed file names
