+++ packages/core/src/types/v3/base-credit-results.ts
