+++ packages/core/src/types/v3/people/get-combined-credits.ts
@@ -1,4 +1,4 @@
-import {PeopleEndpointBaseCrewGetMovieGetCombinedCredit, PeopleEndpointBaseCastGetMovieGetCombinedCredit} from "./bast-cast-crew-credits";
+import {PeopleBaseCrewGetMovieGetCombinedCredit, PeopleBaseCastGetMovieGetCombinedCredit} from "./base-cast-crew-credits";
 
 export interface PeopleGetCombinedCreditsResponse {
   cast: PeopleGetCombinedCreditsCast[];
@@ -6,8 +6,8 @@ export interface PeopleGetCombinedCreditsResponse {
   id: number;
 }
 
-export type PeopleGetCombinedCreditsCast = PeopleEndpointBaseCastGetMovieGetCombinedCredit & {media_type: string};
-export type PeopleGetCombinedCreditsCrew =  PeopleEndpointBaseCrewGetMovieGetCombinedCredit & {media_type: string};
+export type PeopleGetCombinedCreditsCast = PeopleBaseCastGetMovieGetCombinedCredit & {media_type: "movie" | "tv"};
+export type PeopleGetCombinedCreditsCrew =  PeopleBaseCrewGetMovieGetCombinedCredit & {media_type: "movie" | "tv"};
 
 export interface PeopleGetCombinedCreditsParams {
   language?: string;
