+++ packages/core/src/types/v3/tv/get-credits.ts
@@ -1,4 +1,4 @@
-import  {BaseCrewCredit, BaseCastCredit} from "../base-cast-crew-results";
+import  {BaseCrewCredit, BaseCastCredit} from "../base-credit-results";
 
 export interface TVGetCreditsResponse {
   cast: TVGetCreditsCast[];
